requirements.txt
click==8.0.1
cycler==0.10.0
decorator==4.4.2
Flask==2.0.1
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
